server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.4.0
httpx>=0.25.0
langchain>=0.0.350
langchain-openai>=0.0.1
python-dotenv>=1.0.0
